backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1

# LangChain and LangGraph
langchain==0.3.3
langgraph==0.2.38
langchain-openai==0.2.3
langchain-community==0.3.2

# HTTP client for external APIs
httpx==0.27.0

# Tavily search
tavily-python==0.5.0

# Observability (optional)
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
